handle finishing download tasks
